app/requirements.txt
Flask
mysql-connector
pyyaml
requests
python-dotenv
ciso8601